pyproject.toml
[tool.poetry]
name = "data-eng-test"
version = "0.1.0"
description = "Teste prático de implementação de uma pipeline ETL para transferência de dados entre bancos de dados."
authors = ["Nuno de Paula <nunodepaula@gmail.com>"]
readme = "README.md"
packages = [{ include = "etl" }, { include = "db_fonte" }]

[tool.poetry.scripts]
popular_fonte = "db_fonte.populate:populate"
etl = "etl.invoker:execute"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
httpx = "^0.27.2"
SQLAlchemy = "^2.0.36"
fastapi = "^0.115.3"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
numpy = "^2.1.2"
uvicorn = "^0.32.0"
pydantic = "^2.9.2"
argparse = "^1.4.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
ruff = "^0.7.0"
mypy = "^1.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["ALL"]

ignore = [
    "DTZ007", # Por simplicidade, usou-se timestamps e datetimes sem especificação de timezone.
]

pydocstyle.convention = "google"
